Moved DateTime & Timespan construction to the 'Extensions' namespace
