+++ Tests/Shared.Specs/DateTimeAssertionSpecs.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using FluentAssertions.Extensions;
 using FluentAssertions.Primitives;
 using Xunit;
 using Xunit.Sdk;
