+++ Tests/Shared.Specs/DateTimeOffsetAssertionSpecs.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FluentAssertions.Common;
+using FluentAssertions.Extensions;
 using FluentAssertions.Primitives;
 using Xunit;
 using Xunit.Sdk;
