+++ Tests/Shared.Specs/DateTimeOffsetValueFormatterSpecs.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+
+using FluentAssertions.Extensions;
 using FluentAssertions.Formatting;
 using Xunit;
