+++ Tests/Shared.Specs/ExtensibilityRelatedEquivalencySpecs.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using FluentAssertions.Equivalency;
+using FluentAssertions.Extensions;
 using Xunit;
 using Xunit.Sdk;
