+++ Tests/Shared.Specs/ReferenceTypeAssertionsSpecs.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 
 using FluentAssertions.Common;
+using FluentAssertions.Extensions;
 using FluentAssertions.Formatting;
 using Xunit;
 using Xunit.Sdk;
