+++ Tests/Shared.Specs/SimpleTimeSpanAssertionSpecs.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 
 using FluentAssertions.Common;
+using FluentAssertions.Extensions;
 using Xunit;
 using Xunit.Sdk;
