+++ Tests/Shared.Specs/TimeSpanConversionExtensionSpecs.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-
 using FluentAssertions.Common;
+using FluentAssertions.Extensions;
 using Xunit;
 
 namespace FluentAssertions.Specs
